Updated config for updated dependencies
