+++ .browserslistrc
@@ -1,2 +1,1 @@
 last 2 versions
-ie >= 9
