+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "traitify-widgets",
-  "version": "1.10.2",
+  "version": "1.10.3-alpha.0",
   "description": "Traitiy Widgets",
   "repository": "github:traitify/traitify-widgets",
   "main": "./build/traitify.js",
@@ -8,12 +8,11 @@
     "build"
   ],
   "scripts": {
-    "build": "cross-env NODE_ENV=production webpack --progress",
-    "build:compatibility": "cross-env MODE=compatibility NODE_ENV=production webpack  --progress",
-    "dev": "cross-env MODE=browser webpack serve",
-    "dev:compatibility": "cross-env MODE=compatibility webpack serve",
-    "dev:ie": "cross-env MODE=ie webpack serve",
-    "dev:secure": "cross-env MODE=browser webpack serve --https",
+    "build": "webpack --node-env=production --progress",
+    "build:compatibility": "webpack --env mode=compatibility --node-env=production --progress",
+    "dev": "webpack serve --env mode=browser",
+    "dev:compatibility": "webpack serve --env mode=compatibility",
+    "dev:secure": "webpack serve --env mode=browser --https",
     "lint": "eslint {src,test}",
     "lint:fix": "eslint {src,test} --fix",
     "lint:watch": "esw -w {src,test} --color",
@@ -56,13 +55,11 @@
     "@babel/preset-env": "^7.13.9",
     "@babel/register": "^7.13.8",
     "@babel/runtime": "^7.13.9",
-    "@webpack-cli/serve": "^1.3.0",
     "autoprefixer": "^10.2.5",
     "babel-jest": "^26.6.3",
     "babel-loader": "^8.2.2",
     "babel-plugin-jsx-pragmatic": "^1.0.2",
     "babel-preset-airbnb": "^5.0.0",
-    "cross-env": "^7.0.3",
     "css-loader": "^5.1.1",
     "eslint": "^7.21.0",
     "eslint-config-airbnb": "^18.2.1",
@@ -88,9 +85,9 @@
     "sass": "^1.32.8",
     "sass-loader": "^10.1.1",
     "style-loader": "^2.0.0",
-    "webpack": "^5.24.4",
-    "webpack-cli": "^4.5.0",
-    "webpack-dev-server": "^3.11.2"
+    "webpack": "^5.52.1",
+    "webpack-cli": "^4.8.0",
+    "webpack-dev-server": "^4.2.0"
   },
   "devDependenciesNotes": {
     "eslint-plugin-jest": "next version depends on a package that adds a peer-dependency of typescript",
@@ -104,15 +101,12 @@
     "graphql-obj2arg": "^0.1.0",
     "highcharts": "^9.1.0",
     "highcharts-react-official": "^3.0.0",
+    "markdown-to-jsx": "^7.1.3",
     "prop-types": "^15.7.2",
-    "react-markdown": "^4.3.1",
     "regenerator-runtime": "^0.13.7",
-    "uuid": "^3.4.0"
+    "uuid": "^8.3.2"
   },
-  "dependenciesNotes": {
-    "react-markdown": "Version 5+ are incompatible with IE 11",
-    "uuid": "Version 4+ are incompatible with IE 9/10"
-  },
+  "dependenciesNotes": {},
   "peerDependencies": {
     "react": "^16.8.0 || ^17.0.0",
     "react-dom": "^16.8.0 || ^17.0.0"
