+++ public/index.html
@@ -123,6 +123,7 @@
           .allowBack()
           .allowFullscreen()
           .allowHeaders()
+          .allowInstructions()
           .assessmentID(assessmentID)
           .locale(locale)
           .perspective(perspective)
